finish fetch receive data from server and gen icons on the map
